requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
mido>=1.3.0
python-rtmidi>=1.5.0  # Required for mido to work with virtual ports 